Baby database in Excel file
I have selected a variety of cases with different tag combinations, both major tags and their subtypes. It is nice if Radovan could look at it and let me know if any changes in formatting or organizing these data are needed. The leftmost column provides a link to the corresponding page of the construction in the Swedish interface, to show how it looks like now. Note that I have included extra columns (extra regarding to the template uploaded on GitHub), one column per semantic tag, to the right end of the table. These are not the all 52 major semantic tags but only those that account for those constructions included into the baby database. I have also added a column for Structure in UD that was missing in the template. The column FORM is renamed for MORPHOLOGY (FORM).